+++ src/reporter/query_NTNE.py
@@ -48,6 +48,11 @@ def query_NTNE(limit=10000,
     if entities:
         res = []
         for entity in entities:
+            entity['entityId'] = entity['id']
+            entity['entityType'] = entity['type']
+            entity['index'] = entity['index'][0]
+            del entity['id']
+            del entity['type']
             res.append(entity)
         logging.getLogger(__name__).info("Query processed successfully")
         return res